% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{distil}
\alias{distil}
\title{Filter and remove the variables that were filtered on}
\usage{
distil(.data, ..., .by = NULL, .preserve = FALSE)
}
\arguments{
\item{.preserve}{}
}
\description{
Filter and remove the variables that were filtered on
}
