% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{if_not_na}
\alias{if_not_na}
\title{Element selection conditional to NA identity
Copy from above}
\usage{
if_not_na(x, do)
}
\arguments{
\item{do}{}
}
\description{
Element selection conditional to NA identity
Copy from above
}
