% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{rotate_x_axis_labels}
\alias{rotate_x_axis_labels}
\title{Rotate the x axis labels
Helper function}
\usage{
rotate_x_axis_labels(angle = 45, hjust = 1)
}
\arguments{
\item{hjust}{}
}
\description{
Rotate the x axis labels
Helper function
}
