% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_replace_vector}
\alias{str_replace_vector}
\title{Replace vector of patterns with a vector of new text
Every pattern in the vector is matched to a replacement, so that the matches
for the first pattern in the vector are replaced by the first element of the
replacement vector, and so on.}
\usage{
str_replace_vector(x, pattern, replacement)
}
\arguments{
\item{x}{Input string}

\item{pattern}{A vector of patterns to look for.}

\item{replacement}{A vector of replacement strings}
}
\value{
A string with the pattern replacements performed.
}
\description{
Replace vector of patterns with a vector of new text
Every pattern in the vector is matched to a replacement, so that the matches
for the first pattern in the vector are replaced by the first element of the
replacement vector, and so on.
}
